refactor: move util to internal.common package
