+++ core/src/main/java/com/linecorp/armeria/internal/server/annotation/CompositeResponseConverterFunction.java
@@ -25,7 +25,7 @@ import com.linecorp.armeria.common.HttpHeaders;
 import com.linecorp.armeria.common.HttpResponse;
 import com.linecorp.armeria.common.MediaType;
 import com.linecorp.armeria.common.ResponseEntity;
-import com.linecorp.armeria.common.ResponseEntityUtil;
+import com.linecorp.armeria.internal.common.ResponseEntityUtil;
 import com.linecorp.armeria.common.ResponseHeaders;
 import com.linecorp.armeria.common.annotation.Nullable;
 import com.linecorp.armeria.common.util.SafeCloseable;
